extract theme stuff
